requirements.txt
numpy
pyaudio
pyyaml